requirements.txt
alembic==0.9.9
certifi==2018.4.16
click==6.7
Flask==1.0
Flask-Cors==3.0.4
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
GeoAlchemy2==0.4.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
Pillow==5.1.0
psycopg2-binary==2.7.4
python-dateutil==2.7.2
python-editor==1.0.3
python-http-client==3.0.0
reportlab==3.4.0
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
https://github.com/PnX-SI/UsersHub-authentification-module/archive/1.4.3.zip
